+++ packages/components/src/SegmentedControl/SegmentedControlBase.tsx
@@ -1,0 +1,31 @@
+import React, {
+  CSSProperties,
+  Children,
+  PropsWithChildren,
+  forwardRef,
+} from "react";
+import styles from "./SegmentedControl.module.css";
+
+const SegmentedControlBase = forwardRef<HTMLDivElement, PropsWithChildren>(
+  function SegmentedControlBase({ children }, ref) {
+    const optionCount = Children.count(children);
+
+    return (
+      <div
+        ref={ref}
+        className={styles.container}
+        role="radiogroup"
+        style={
+          {
+            "--segmentedControl--option-count": optionCount,
+          } as CSSProperties
+        }
+      >
+        {children}
+        <span />
+      </div>
+    );
+  },
+);
+
+export { SegmentedControlBase };