+++ packages/components/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.48.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@5.47.1...@jobber/components@5.48.0) (2024-11-14)
+
+
+### Features
+
+* **components:** SegmentedControl ([#1966](https://github.com/GetJobber/atlantis/issues/1966)) ([c403bd1](https://github.com/GetJobber/atlantis/commit/c403bd11b6eebfcaa361e5158770b0c8bb178b96))
+
+
+
+
+
+## [5.47.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@5.47.0...@jobber/components@5.47.1) (2024-11-12)
+
+
+### Bug Fixes
+
+* **components:** Remove unnecessary useRefocusOnActivator from DataListHeaderTile ([#2111](https://github.com/GetJobber/atlantis/issues/2111)) ([7e1a32a](https://github.com/GetJobber/atlantis/commit/7e1a32a41d21259294e8fa3a38a481e6a3ae4efa))
+
+
+
+
+
 # [5.47.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@5.46.1...@jobber/components@5.47.0) (2024-11-05)